rsz: Retire Rebuffer.cc
