+++ src/rsz/src/BufferMove.cc
@@ -30,9 +30,6 @@ using sta::Vertex;
 
 BufferMove::BufferMove(Resizer* resizer) : BaseMove(resizer)
 {
-  for (auto rf_index : RiseFall::rangeIndex()) {
-    arrival_paths_[rf_index] = nullptr;
-  }
 }
 
 bool BufferMove::doMove(const Path* drvr_path,
